update: move token grant tests to new parent folder
