+++ tests/AuthServer.Tests.UnitTest/TokenByGrant/ClientCredentialsGrant/ClientCredentialsRequestProcessorTest.cs
@@ -8,7 +8,7 @@ using AuthServer.TokenByGrant.ClientCredentialsGrant;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit.Abstractions;
 
-namespace AuthServer.Tests.UnitTest.ClientCredentialsGrant;
+namespace AuthServer.Tests.UnitTest.TokenByGrant.ClientCredentialsGrant;
 public class ClientCredentialsRequestProcessorTest : BaseUnitTest
 {
     public ClientCredentialsRequestProcessorTest(ITestOutputHelper outputHelper)
